# public/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentStamp API",
    "version": "2.3.0",
    "description": "AgentStamp provides cryptographic identity certificates (stamps), a searchable public registry, reputation scores, cross-protocol passports, verification badges, and a Wishing Well marketplace for AI agents. Free tier available. Paid endpoints use x402 micropayments (USDC on Base and Solana).",
    "contact": {
      "name": "AgentStamp",
      "url": "https://agentstamp.org"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://agentstamp.org",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/v1/stamp/mint/free": {
      "post": {
        "operationId": "mintFreeStamp",
        "summary": "Mint a free 7-day identity stamp",
        "description": "Mint a free identity certificate valid for 7 days. One free stamp per wallet per 7 days.",
        "tags": ["Stamp"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalletAddressBody"
              }
            }
          }
        },
        "parameters": [
          { "$ref": "#/components/parameters/XWalletAddress" }
        ],
        "responses": {
          "201": {
            "description": "Stamp minted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FreeStampMintResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing wallet address",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "429": {
            "description": "Cooldown active — one free stamp per wallet per 7 days",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CooldownErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/stamp/mint/{tier}": {
      "post": {
        "operationId": "mintStamp",
        "summary": "Mint a paid identity stamp",
        "description": "Mint a paid identity certificate. Requires x402 USDC payment. Bronze ($0.001, 24h), Silver ($0.005, 7d), Gold ($0.01, 30d).",
        "tags": ["Stamp"],
        "x-pricing": {
          "bronze": { "price": "$0.001", "currency": "USDC", "networks": ["Base", "Solana"], "protocol": "x402" },
          "silver": { "price": "$0.005", "currency": "USDC", "networks": ["Base", "Solana"], "protocol": "x402" },
          "gold": { "price": "$0.01", "currency": "USDC", "networks": ["Base", "Solana"], "protocol": "x402" }
        },
        "parameters": [
          {
            "name": "tier",
            "in": "path",
            "required": true,
            "description": "Stamp tier determining validity and price",
            "schema": {
              "type": "string",
              "enum": ["bronze", "silver", "gold"]
            }
          },
          { "$ref": "#/components/parameters/XWalletAddress" }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalletAddressBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Stamp minted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StampMintResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid tier or validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Wallet address required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "402": {
            "description": "Payment required — x402 USDC payment not provided or insufficient"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/stamp/verify/{certId}": {
      "get": {
        "operationId": "verifyStamp",
        "summary": "Verify a certificate",
        "description": "Verify the validity and signature of an identity certificate by its ID.",
        "tags": ["Stamp"],
        "parameters": [
          {
            "name": "certId",
            "in": "path",
            "required": true,
            "description": "The stamp/certificate ID to verify",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StampVerifyResponse"
                }
              }
            }
          },
          "404": {
            "description": "Certificate not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/stamp/stats": {
      "get": {
        "operationId": "getStampStats",
        "summary": "Get stamp statistics",
        "description": "Returns aggregate statistics about stamps: total issued, active, by tier, today, and this week.",
        "tags": ["Stamp"],
        "responses": {
          "200": {
            "description": "Stamp statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StampStatsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "XWalletAddress": {
        "name": "x-wallet-address",
        "in": "header",
        "description": "Wallet address (EVM or Solana). Can also be provided as wallet_address in the request body.",
        "schema": { "type": "string" }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": { "type": "boolean", "const": false },
          "error": { "type": "string" }
        }
      },
      "CooldownErrorResponse": {
        "type": "object",
        "required": ["success", "error", "next_available"],
        "properties": {
          "success": { "type": "boolean", "const": false },
          "error": { "type": "string" },
          "next_available": {
            "type": "string",
            "format": "date-time",
            "description": "ISO 8601 timestamp of when the next free action is available"
          }
        }
      },
...
      "WalletAddressBody": {
        "type": "object",
        "properties": {
          "wallet_address": {
            "type": "string",
            "description": "EVM or Solana wallet address"
          }
        }
      },
      "Certificate": {
        "type": "object",
        "properties": {
          "version": { "type": "string", "example": "1.0" },
          "issuer": { "type": "string", "example": "AgentStamp" },
          "subject": { "type": "string", "description": "Wallet address" },
          "tier": {
            "type": "string",
            "enum": ["free", "bronze", "silver", "gold"]
          },
          "paid": { "type": "boolean" },
          "claim": { "type": "string" },
          "issued_at": { "type": "string", "format": "date-time" },
          "expires_at": { "type": "string", "format": "date-time" },
          "stamp_id": { "type": "string" }
        }
      },
      "Stamp": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "tier": {
            "type": "string",
            "enum": ["free", "bronze", "silver", "gold"]
          },
          "wallet_address": { "type": "string" },
          "issued_at": { "type": "string", "format": "date-time" },
          "expires_at": { "type": "string", "format": "date-time" },
          "certificate": { "$ref": "#/components/schemas/Certificate" },
          "signature": { "type": "string", "description": "Ed25519 signature of the certificate" },
          "verify_url": { "type": "string" }
        }
      },
      "FreeStampMintResponse": {
        "type": "object",
        "required": ["success", "stamp"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "stamp": { "$ref": "#/components/schemas/Stamp" },
          "upgrade_info": {
            "type": "object",
            "properties": {
              "message": { "type": "string" },
              "tiers": {
                "type": "object",
                "properties": {
                  "bronze": {
                    "type": "object",
                    "properties": {
                      "price": { "type": "string", "example": "$0.001" },
                      "validity": { "type": "string", "example": "24h" }
                    }
                  },
                  "silver": {
                    "type": "object",
                    "properties": {
                      "price": { "type": "string", "example": "$0.005" },
                      "validity": { "type": "string", "example": "7 days" }
                    }
                  },
                  "gold": {
                    "type": "object",
                    "properties": {
                      "price": { "type": "string", "example": "$0.01" },
                      "validity": { "type": "string", "example": "30 days" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "StampMintResponse": {
        "type": "object",
        "required": ["success", "stamp"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "stamp": { "$ref": "#/components/schemas/Stamp" }
        }
      },
      "StampVerifyResponse": {
        "type": "object",
        "required": ["success", "valid", "expired", "revoked", "stamp", "public_key"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "valid": {
            "type": "boolean",
            "description": "True if signature is valid, not expired, and not revoked"
          },
          "expired": { "type": "boolean" },
          "revoked": { "type": "boolean" },
          "stamp": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "tier": { "type": "string", "enum": ["free", "bronze", "silver", "gold"] },
              "wallet_address": { "type": "string" },
              "issued_at": { "type": "string", "format": "date-time" },
              "expires_at": { "type": "string", "format": "date-time" },
              "certificate": { "$ref": "#/components/schemas/Certificate" }
            }
          },
          "public_key": {
            "type": "string",
            "description": "Ed25519 public key used to verify the signature"
          }
        }
      },
      "StampStatsResponse": {
        "type": "object",
        "required": ["success", "stats"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "stats": {
            "type": "object",
            "properties": {
              "total_issued": { "type": "integer" },
              "active": { "type": "integer" },
              "by_tier": {
                "type": "object",
                "additionalProperties": { "type": "integer" },
                "description": "Count of stamps by tier (e.g. { free: 10, bronze: 5, gold: 2 })"
              },
              "today": { "type": "integer" },
              "this_week": { "type": "integer" }
            }
          }
        }
      },
...
    }
  }
}